Fixing the build for upload-failure lambda
