+++ Dockerfile.upload_failures
@@ -1,6 +1,7 @@
 FROM public.ecr.aws/lambda/python:3.8
 
 COPY backend/corpora/upload_failure .
+
 RUN pip3 install -r requirements.txt
 
 COPY backend/__init__.py ./backend/__init__.py
