+++ Makefile
@@ -233,6 +233,11 @@ local-uploadfailure: .env.ecr ## Run the upload failure lambda with a dataset id
 	docker-compose $(COMPOSE_OPTS) up -d upload_failures
 	curl -v -XPOST "http://127.0.0.1:9000/2015-03-31/functions/function/invocations" -d '{"dataset_id": "$(DATASET_ID)", "error": {"Cause": "$(CAUSE)"}}'
 
+.PHONY: local-uploadsuccess
+local-uploadsuccess: .env.ecr ## Run the upload success lambda with a dataset id and cause
+	docker-compose $(COMPOSE_OPTS) up -d upload_success
+	curl -v -XPOST "http://127.0.0.1:9001/2015-03-31/functions/function/invocations" -d '{"dataset_id": "$(DATASET_ID)"}'
+
 .PHONY: local-cxguser-cookie
 local-cxguser-cookie: ## Get cxguser-cookie
 	docker-compose $(COMPOSE_OPTS) run --rm backend bash -c "cd /single-cell-data-portal && python login.py"
