+++ docker-compose.yml
@@ -83,7 +83,6 @@ services:
       - "9000:8080"
     volumes:
       - ./backend/corpora/upload_failures:/var/task
-      - ./backend/corpora/lambdas/upload_failures:/var/task/backend/corpora/upload_failures
       - ./backend/corpora/common:/var/task/backend/corpora/common
       - ./backend/corpora/dataset_processing:/var/task/backend/corpora/dataset_processing
     environment:
