+++ tests/unit/backend/corpora/common/entities/datasets/test_update.py
@@ -6,7 +6,7 @@ from backend.corpora.common.corpora_orm import (
 )
 from backend.corpora.common.entities import Dataset
 from backend.corpora.common.utils.db_helpers import processing_status_updater
-from backend.corpora.upload_failures.upload import update_dataset_processing_status_to_failed
+from backend.corpora.upload_failures.app import update_dataset_processing_status_to_failed
 from tests.unit.backend.corpora.common.entities.datasets import TestDataset
 from tests.unit.backend.utils import BogusProcessingStatusParams, BogusDatasetParams
